+++ api/src/main/java/dev/nolij/zume/api/util/v0/package-info.java
@@ -1,4 +1,0 @@
-@ApiStatus.Experimental
-package dev.nolij.zume.api.util.v0;
-
-import org.jetbrains.annotations.ApiStatus;