+++ modern/src/main/java/dev/nolij/zume/mixin/modern/GameRendererMixin.java
@@ -1,7 +1,7 @@
 package dev.nolij.zume.mixin.modern;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import dev.nolij.zume.api.platform.v0.ZumeAPI;
+import dev.nolij.zume.api.platform.v1.ZumeAPI;
 import net.minecraft.client.renderer.GameRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;