+++ modern/src/main/java/dev/nolij/zume/mixin/modern/MouseHandlerMixin.java
@@ -2,7 +2,7 @@ package dev.nolij.zume.mixin.modern;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
-import dev.nolij.zume.api.platform.v0.ZumeAPI;
+import dev.nolij.zume.api.platform.v1.ZumeAPI;
 import net.minecraft.client.MouseHandler;
 import net.minecraft.world.entity.player.Inventory;
 import org.spongepowered.asm.mixin.Dynamic;