+++ modern/src/main/java/dev/nolij/zume/modern/integration/ModernZumeConfigScreen.java
@@ -1,6 +1,6 @@
 package dev.nolij.zume.modern.integration;
 
-import dev.nolij.zume.api.platform.v0.ZumeAPI;
+import dev.nolij.zume.api.platform.v1.ZumeAPI;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;