+++ primitive/src/main/java/dev/nolij/zume/mixin/primitive/GameRendererMixin.java
@@ -2,7 +2,7 @@ package dev.nolij.zume.mixin.primitive;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import dev.nolij.zume.api.platform.v0.ZumeAPI;
+import dev.nolij.zume.api.platform.v1.ZumeAPI;
 import net.minecraft.class_555;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;