+++ vintage/src/main/java/dev/nolij/zume/mixin/vintage/EntityRendererMixin.java
@@ -1,7 +1,7 @@
 package dev.nolij.zume.mixin.vintage;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import dev.nolij.zume.api.platform.v0.ZumeAPI;
+import dev.nolij.zume.api.platform.v1.ZumeAPI;
 import net.minecraft.client.renderer.EntityRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;